+++ lib/Github/Api/CurrentUser/PublicKeys.php
@@ -9,7 +9,7 @@ use Github\Exception\MissingArgumentException;
  * @link   http://developer.github.com/v3/users/keys/
  * @author Joseph Bielawski <stloyd@gmail.com>
  */
-class DeployKeys extends AbstractApi
+class PublicKeys extends AbstractApi
 {
     /**
      * List deploy keys for the authenticated user.